+++ packages/context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/context
 
+## 0.29.5
+
+### Patch Changes
+
+- @tangleai/models@0.29.5
+
 ## 0.29.4
 
 ### Patch Changes