+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/core
 
+## 0.29.5
+
+### Patch Changes
+
+- @tangleai/context@0.29.5
+
 ## 0.29.4
 
 ### Patch Changes