+++ packages/jaren/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tangleai/jaren
 
+## 0.29.5
+
+### Patch Changes
+
+- @tangleai/agents@0.29.5
+  - @tangleai/context@0.29.5
+  - @tangleai/models@0.29.5
+
 ## 0.29.4
 
 ### Patch Changes